k8s-provisioner/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==3.2.2

# Celery and Redis
celery==5.3.6
redis==5.0.1
flower==2.0.1

# SSH and Ansible
paramiko==3.4.0
ansible==9.2.0
ansible-runner==2.3.4

# Security and encryption
cryptography==42.0.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Monitoring (optional)
prometheus-client==0.19.0

# WebSocket support
python-socketio==5.11.0

# Rate limiting
slowapi==0.1.9

# Date and time
python-dateutil==2.8.2
pytz==2024.1
